4.15-4.23

+++ part4/blog/controllers/blogController.js
@@ -1,8 +1,10 @@
 const Blog = require('../models/blog');
 const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
 const blogRouter = require('express').Router();
 
+//+ get
 blogRouter.get('/', (req, response) => {
   Blog.find({})
     .populate('user', { username: 1 })
@@ -11,19 +13,35 @@ blogRouter.get('/', (req, response) => {
     });
 });
 
+const getTokenFrom = (request) => {
+  const authorization = request.get('authorization');
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7);
+  }
+  return null;
+};
+
+//$ Post new
 blogRouter.post('/', async (req, res) => {
+  const token = req.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const user = await User.findById(decodedToken.id);
+
   if (!req.body.likes) {
     req.body.likes = 0;
   }
 
-  const user = await User.findById(req.body.user);
+  const currentUser = await User.findById(req.body.user);
 
   const blog = new Blog({
     title: req.body.title,
     author: req.body.author,
     url: req.body.url,
     likes: req.body.likes,
-    user: user._id,
+    user: currentUser._id,
   });
 
   const savedBlog = await blog.save();
@@ -32,14 +50,24 @@ blogRouter.post('/', async (req, res) => {
   await user.save();
 });
 
+//! delete
 blogRouter.post('/:id/delete', async (req, res) => {
+  const token = req.token;
+  const find = Blog.findById(req.params.id);
+  if (find.user.toString() !== token.toString()) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
   const remove = await Blog.findByIdAndRemove(req.params.id);
   res.status(204).json(remove);
 });
 
+//# update
 blogRouter.post('/:id/like', async function (req, res, next) {
   const before = await Blog.findOne({ id: req.params.id }).exec();
-  let blog = await Blog.findOneAndUpdate({ id: req.params.id }, { likes: before.likes + 1 });
+  let blog = await Blog.findOneAndUpdate(
+    { id: req.params.id },
+    { likes: before.likes + 1 }
+  );
   res.status(204).json(blog);
 });
